// packages/version-1/error/types.ts
// Error-log information
export interface ErrorLogInfo {
	// Type of error
	type: ErrorTypeEnum;
	// Error file
	file: string;
	// Error name
	name?: string;
	// Error message
	message: string;
	// Error stack
	stack?: string;
	// Error detail
	detail: string;
	// Error url
	url: string;
	// Error time
	time?: string;
}

export enum ErrorTypeEnum {
	VUE = 'vue',
	SCRIPT = 'script',
	RESOURCE = 'resource',
	AJAX = 'ajax',
	PROMISE = 'promise',
}